Re-style new topic page with tailwind
