--- src/pages/NewTopic/NewTopic.scss
@@ -1,25 +1,0 @@
-@import "../../utils";
-
-.new-topic {
-    nav {
-        justify-content: space-between;
-    }
-
-    .content {
-        @include height($points: 92);
-        @include centered-flex-container;
-        flex-direction: column;
-        gap: 1em;
-
-        input {
-            padding: 0 0.5em;
-            height: 3rem;
-            width: 70%;
-            font-size: 1.2rem;
-            text-align: center;
-            outline: none;
-            border: 1px solid $black;
-            border-radius: 0.5em;
-        }
-    }
-}
